Code cleanups. Signed/unsigned comparisons

+++ exult/trunk/gamedat.cc
@@ -249,7 +249,7 @@ void Game_window::read_save_names
 	(
 	)
 	{
-	for (int i = 0; i < sizeof(save_names)/sizeof(save_names[0]); i++)
+	for (size_t i = 0; i < sizeof(save_names)/sizeof(save_names[0]); i++)
 		{
 		char fname[50];		// Set up name.
 		sprintf(fname, SAVENAME, i);
